% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_utils.R
\name{get_checkout_doubles}
\alias{get_checkout_doubles}
\title{Get All Possible Doubles}
\usage{
get_checkout_doubles()
}
\value{
Integer vector of valid double scores.
}
\description{
Returns all valid checkout doubles (for double-out games).
}
\examples{
get_checkout_doubles()

}
