% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_utils.R
\name{get_valid_dart_scores}
\alias{get_valid_dart_scores}
\title{Get All Valid Dart Scores}
\usage{
get_valid_dart_scores()
}
\value{
Integer vector of valid dart scores.
}
\description{
Returns a vector of all possible scores achievable with a single dart.
}
\examples{
get_valid_dart_scores()

}
