% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_utils.R
\name{max_turn_score}
\alias{max_turn_score}
\title{Calculate Maximum Possible Turn Score}
\usage{
max_turn_score(num_darts = 3L)
}
\arguments{
\item{num_darts}{Integer. Number of darts (1-3). Default is 3.}
}
\value{
Integer maximum score.
}
\description{
Returns the maximum score achievable in a single turn (180 with 3 darts).
}
\examples{
max_turn_score()      # 180
max_turn_score(1)     # 60
max_turn_score(2)     # 120

}
